% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{make_DT_table}
\alias{make_DT_table}
\title{make_DT_table}
\usage{
make_DT_table(
  DF,
  editable = F,
  selection = "single",
  paging = TRUE,
  scrollY = F,
  searching = T
)
}
\description{
make_DT_table
}
