% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R, R/tables.R
\name{make_table1}
\alias{make_table1}
\title{make_table1}
\usage{
make_table1(df_table, group = "no_choice", variables, render.missing = F)

make_table1(df_table, group = "no_choice", variables, render.missing = F)
}
\description{
make_table1

make_table1
}
