% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials.R
\name{trial_diagram}
\alias{trial_diagram}
\title{Generate Trial Diagram}
\usage{
trial_diagram(
  trial,
  static = FALSE,
  render = TRUE,
  include_drugs = TRUE,
  duplicate_drugs = FALSE,
  hierarchical = TRUE,
  direction = "LR",
  zoomView = TRUE
)
}
\description{
Generate Trial Diagram
}
