% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials.R
\name{trial_diagram2}
\alias{trial_diagram2}
\title{Generate Trial Diagram from NCT}
\usage{
trial_diagram2(
  nct_id,
  static = FALSE,
  render = TRUE,
  include_drugs = TRUE,
  duplicate_drugs = FALSE,
  hierarchical = TRUE,
  direction = "LR",
  zoomView = TRUE
)
}
\description{
Generate Trial Diagram from NCT
}
